requirements.txt
# torch==1.13.1
# torchvision==0.14.1
# pip install torch==1.13.1+cu117 torchvision==0.14.1+cu117 torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cu117
albumentations==1.3.0
opencv-python==4.6.0.66
imageio==2.9.0
numpy==1.23.1
imageio-ffmpeg==0.4.2
pytorch-lightning==1.4.2
omegaconf==2.1.1
test-tube>=0.7.5
streamlit==1.12.1
einops==0.3.0
transformers==4.19.2
webdataset==0.2.5
kornia==0.6
open_clip_torch==2.0.2
invisible-watermark>=0.1.5
streamlit-drawable-canvas==0.8.0
torchmetrics==0.6.0
xformers
triton
matplotlib
wandb
pillow
# -e git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
# -e git+https://github.com/openai/CLIP.git@main#egg=clip
# -e .


# additional requirements
timm
matplotlib
open_clip_torch
requests==2.27.1
urllib3==1.25.11
